RAG/requirements.txt
flask
openai
pandas
python-dotenv
google-cloud-storage
hugchat
